+++ src/main/java/com/example/onlyone/domain/user/service/UserService.java
@@ -3,8 +3,8 @@ package com.example.onlyone.domain.user.service;
 import com.example.onlyone.domain.interest.entity.Category;
 import com.example.onlyone.domain.interest.entity.Interest;
 import com.example.onlyone.domain.interest.repository.InterestRepository;
-import com.example.onlyone.domain.settlement.dto.response.MySettlementDto;
-import com.example.onlyone.domain.settlement.dto.response.MySettlementResponseDto;
+import com.example.onlyone.domain.user.dto.response.MySettlementDto;
+import com.example.onlyone.domain.user.dto.response.MySettlementResponseDto;
 import com.example.onlyone.domain.settlement.repository.UserSettlementRepository;
 import com.example.onlyone.domain.user.dto.request.ProfileUpdateRequestDto;
 import com.example.onlyone.domain.user.dto.request.SignupRequestDto;
@@ -18,7 +18,6 @@ import com.example.onlyone.domain.user.repository.UserInterestRepository;
 import com.example.onlyone.domain.user.repository.UserRepository;
 import com.example.onlyone.domain.wallet.entity.Wallet;
 import com.example.onlyone.domain.wallet.repository.WalletRepository;
-import com.example.onlyone.domain.wallet.service.WalletService;
 import com.example.onlyone.global.exception.CustomException;
 import com.example.onlyone.global.exception.ErrorCode;
 import io.jsonwebtoken.Jwts;
